requirements.txt
akshare
requests
openai
python-dotenv
chinese-calendar
pandas
beautifulsoup4
html5lib
lxml
matplotlib
flask
flask-cors
flask-apscheduler
